Function to plot sex ratio

Function to plot sex ratio

Usage

PlotSexRatio.fn(
  dir,
  dat,
  data.type = "length",
  main = NULL,
  circleSize = 0.1,
  dopng = lifecycle::deprecated(),
  ...
)

Arguments

dir

directory location for saving the png

dat

data object

data.type

"length" or "age"

main

ame that will be used to name the saved png

circleSize

circle size

dopng

Deprecated with nwfscSurvey 2.1 because providing a non-NULL value to dir can serve the same purpose as dopng = TRUE without the potential for errors when dopng = TRUE and dir = NULL. Thus, users no longer have to specify dopng to save the plot as a png.

...

Additional arguments for the plots

Author

Allan Hicks and Chantel Wetzel