A data set for determining strata areas (hectares)

Data necessary for determining the amount of area a strata covers to expand estimates of biomass to be absolute. The SA3 data set was updated in March of 2021 to allow for splitting areas at the 40.16667, a common management line. The data includes the following columns:

OBJECT_ID

MIN_DEPTH_M

MAX_DEPTH_M

MIN_LAT_DD

MAX_LAT_DD

AREA_HECTARES

INPFC_AREA

Usage

SA3_v2021.1

Format

A data frame.

Author

Curt Whitmire