Calculates proportion of age at length and reformats into SS format Uses raw numbers at length, assuming that is a random sample conditioned on length and sex. To use expanded numbers (up to strata areas), set raw=F Only sex codes 1 and 2 and puts males and females on separate lines because the age@L is conditioned on sex (a sample of females of length 25cm, for example) Gender=1: females only. Male values ignored Gender=2: males only. Female values ignored. lgthBins is either the interval between length bins or the actual length bins note that 0 and Inf are tacked on the ends to account for lengths and ages outside the interval. You may want to add these in to first and last bin. I assume all fish are sexed for age data, thus do not apply sex ratios for unsexed fish

Calculates proportion of age at length and reformats into SS format Uses raw numbers at length, assuming that is a random sample conditioned on length and sex. To use expanded numbers (up to strata areas), set raw=F Only sex codes 1 and 2 and puts males and females on separate lines because the age@L is conditioned on sex (a sample of females of length 25cm, for example) Gender=1: females only. Male values ignored Gender=2: males only. Female values ignored. lgthBins is either the interval between length bins or the actual length bins note that 0 and Inf are tacked on the ends to account for lengths and ages outside the interval. You may want to add these in to first and last bin. I assume all fish are sexed for age data, thus do not apply sex ratios for unsexed fish

Usage

SurveyAgeAtLen.fn(
  dir = NULL,
  datAL,
  datTows,
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df = NULL,
  lgthBins = 1,
  ageBins = 1,
  sex = 3,
  SSout = TRUE,
  meanRatioMethod = TRUE,
  raw = TRUE,
  NAs2zero = TRUE,
  month = "Enter Month",
  fleet = "Enter Fleet",
  partition = 0,
  ageErr = "Enter Age Error",
  returnSamps = FALSE,
  printfolder = "forSS",
  verbose = TRUE
)

Arguments

dir

directory this is where the output files will be saved

datAL

the biological data frame exctrated from the data warehouse using the PullBio.fn

datTows

the catch data frame extracted from the data warehouse using the PullCatch.fn

strat.vars

the variables used define the stratas. Defaul is bottom depth and latitudes.

strat.df

the created strata matrix with the calculated areas by the createStrataDF.fn function

lgthBins

length bins

ageBins

age bins

sex

sex (0, 1, 2, 3) sex value for Stock Synthesis

SSout

TRUE/FALSE return comps formatted for SS or in a raw form

meanRatioMethod

TRUE/FALSE

raw

TRUE/FALSE, input to define whether or not to expand numbers in the csv file (column header "NumF" and "NumM")

NAs2zero

change NAs to zeros

month

month

fleet

fleet number

partition

partition for Stock Synthesis

ageErr

age error value for Stock Synthesis

returnSamps

TRUE/FALSE stops the function after the sample size is calculated

printfolder

folder where the length comps will be saved

verbose

A logical that specifies if you want to print messages and warnings to the console. The default is TRUE.

See also

Author

Allan Hicks and Chantel Wetzel