This function plots frequency data as bubble plots You may want to change all zeros to NA's so that those observations are not plotted. If you don't then set zero2NAs=F

This function plots frequency data as bubble plots You may want to change all zeros to NA's so that those observations are not plotted. If you don't then set zero2NAs=F

Usage

PlotFreqData.fn(
  dir = NULL,
  dat,
  inch = 0.15,
  ylab = "Bins",
  xlab = "Year",
  zero2NAs = T,
  main = NULL,
  xlim = NULL,
  ymax = NULL,
  dopng = lifecycle::deprecated(),
  w = 7,
  h = 7,
  ...
)

Arguments

dir

directory to save files to

dat

object created by SS3LF.fn or SS3AF.fn

inch

input to the symbols plot: TRUE, FALSE or a positive number.

ylab

y-axis text label

xlab

x-axis text label

zero2NAs

T/F change 0 values to NA

main

main plot text

xlim

x-limit values

ymax

Value used to truncate y-axis, defaults to NULL

dopng

Deprecated with nwfscSurvey 2.1 because providing a non-NULL value to dir can serve the same purpose as dopng = TRUE without the potential for errors when dopng = TRUE and dir = NULL. Thus, users no longer have to specify dopng to save the plot as a png.

w

Numeric figure width, defaults to 7

h

Numeric figure height, defaults to 7

...

Additional arguments for the plots

Author

Allan Hicks and Chantel Wetzel