Function to plot sex ratio by strata

Function to plot sex ratio by strata

Usage

PlotSexRatioStrata.fn(
  dir = NULL,
  dat,
  type = "length",
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df = NULL,
  circleSize = 0.05,
  dopng = lifecycle::deprecated(),
  ...
)

Arguments

dir

directory location for saving the png

dat

data object

type

length/age which data type to use

strat.vars

the parameters to stratify the data

strat.df

the created strata matrix with the calculated areas by the createStrataDF.fn function

circleSize

circle size

dopng

Deprecated with nwfscSurvey 2.1 because providing a non-NULL value to dir can serve the same purpose as dopng = TRUE without the potential for errors when dopng = TRUE and dir = NULL. Thus, users no longer have to specify dopng to save the plot as a png.

...

Additional arguments for the plots

See also

Author

Allan Hicks and Chantel Wetzel