This function plots frequency data as bubble plots

This function plots frequency data as bubble plots

Usage

plot_comps(
  data,
  dir = NULL,
  add_save_name = NULL,
  plot = 1:2,
  add_0_ylim = TRUE,
  width = 10,
  height = 7
)

Arguments

data

Data file object created by SS3LF.fn or SS3AF.fn

dir

Directory to save files to

add_save_name

Option to add text to a saved figure name. This option can be useful if creating plots across multiple species and saving them into a single folder. Default NULL.

plot

A vector of integers to specify which plots you would like. The default is to print both figures.

bubble plot of length/age composition data by year and sex

distribution by year of length/age compositions data similar to r4ss figure

add_0_ylim

TRUE/FALSE, option to specificy if the y-axis should start at 0 or the minimum bin in the composition data file. Default TRUE.

width

Numeric figure width in inches, defaults to 10

height

Numeric figure height in inches, defaults to 7

Author

Chantel Wetzel

Examples

if (FALSE)  plot_comps(data = LFs)