Plot showing presence and absence per haul by depth, latitude, or sex bin

Uses the data.frame of data values extracted from the data warehouse by either PullCatch.fn() or PullBio.fn() to make a table of proportion by depth or latitude bins. If biological data are passed to the function then there is a single option to plot the sex ratio by depth.

Usage

plot_proportion(
  data,
  dim = c("depth", "lat", "sex"),
  dir = NULL,
  plot_type = c("proportion", "total"),
  depth_min = 50,
  depth_max = NULL,
  depth_bin_width = 25,
  lat_min = 32,
  lat_max = 49,
  lat_bin_width = 1,
  add_range_to_main = TRUE,
  xlab = NULL
)

Arguments

data

data.frame containing data per haul created by PullCatch.fn() or biological data created by PullBio.fn() where dim must be dim = "sex".

dim

Dimension of interest, either "depth", "lat", or "sex".

dir

The directory name to save figures to where they will be save via file.path(dir, "plots").

plot_type

Two options area available, "proportion" or "total" where the default, "proportion", plots the proportion by depth/latitude with equal bar widths and the "total" option plots the numbers by depth/latitude with the bar widths in relation to sampling by depth/latitude.

depth_min

Minimum depth (in meters).

depth_max

Maximum depth (in meters). A NULL value will cause the function to automatically set depth_max to the multiple of depth_bin_width beyond the 99.9 percentile of the observations.

depth_bin_width

Width of each depth bin (in meters).

lat_min

Minimum latitude (in decimal degrees).

lat_max

Maximum latitude (in decimal degrees).

lat_bin_width

Width of each latitude bin (in decimal degrees).

add_range_to_main

Add the range of latitude or depth by which the data are filtered.

xlab

Label for x-axis. A NULL value will cause the function to choose either "Depth (m)" or "Latitude (°N)".

Author

Ian G. Taylor and Chantel Wetzel

Examples

if (FALSE) {
# load WCGBTS data data
data.WCGBTS.ling <- nwfscSurvey::PullCatch.fn(
  Name = "lingcod",
  SurveyName = "NWFSC.Combo"
)
bio.WCGBTS.ling <- nwfscSurvey::PullBio.fn(
  Name = "lingcod",
  SurveyName = "NWFSC.Combo"
)
plot_proportion(data = data.WCGBTS.ling, dim = "lat")
plot_proportion(data = data.WCGBTS.ling, dim = "depth")
plot_proportion(data = bio.WCGBTS.ling, dim = "sex")
}